+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - CI-342 - Promoted similar projects on `CapitalInvestOpportunityPage` are now randomised
 - CI-344 - `Similar projects` title in `CapitalInvestOpportunityPage` is now hard coded
 - CI-332 - Sass changes for mobile design for `CapitalInvestOpportunityListingPage`
+- CI-323 - Add "Why choose the UK" page to "About the UK" section
 - CI-372 - Updated trade CTA on sector pages to go to trade contact form